fix(monitor): Refactor table into component
